mypt-iqc-api/requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cron-schedule==1.1
croniter==1.3.8
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
kafka-python==2.0.2
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.23.4
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
redis==4.3.4
requests==2.28.1
schedule==1.1.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
urllib3==1.26.12
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0
